+++ src/main/java/com/eventitta/common/config/AuditorAwareImpl.java
@@ -1,16 +1,17 @@
 package com.eventitta.common.config;
 
-import com.eventitta.auth.jwt.constants.SecurityConstants;
 import com.eventitta.common.util.SecurityUtil;
 import org.springframework.data.domain.AuditorAware;
 
 import java.util.Optional;
+
+import static com.eventitta.auth.constants.AuthConstants.ANONYMOUS;
 
 public class AuditorAwareImpl implements AuditorAware<String> {
 
     @Override
     public Optional<String> getCurrentAuditor() {
         String userName = SecurityUtil.getCurrentUserName();
-        return SecurityConstants.ANONYMOUS.equals(userName) ? Optional.empty() : Optional.of(userName);
+        return ANONYMOUS.equals(userName) ? Optional.empty() : Optional.of(userName);
     }
 }
